Use Capture api for mouse down/up on windows
This is useful especially for click and drag events like moving a scrollbar. This allows you to continue to receive mouse events even if you accidentally move the cursor outside the window. It will also allow the mouse up event to come through even if it happens outside the window.